イベントのメモ入力ができない不具合を修正
